+++ server/strategy_test.go
@@ -547,6 +547,7 @@ func (d testData) withK8sResourceReady() testData {
 			Annotations: map[string]string{cke.AnnotationResourceRevision: "1"},
 		}
 	}
+	ks.ResourceStatuses["ClusterRole/system:kube-apiserver-to-kubelet"].Annotations[cke.AnnotationResourceRevision] = "2"
 	ks.ResourceStatuses["ClusterRole/system:cluster-dns"].Annotations[cke.AnnotationResourceRevision] = "2"
 	ks.ResourceStatuses["Deployment/kube-system/cluster-dns"].Annotations[cke.AnnotationResourceImage] = cke.CoreDNSImage.Name()
 	ks.ResourceStatuses["Deployment/kube-system/cluster-dns"].Annotations[cke.AnnotationResourceRevision] = "5"