+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project employs a versioning scheme described in [RELEASE.md](RELEASE.md#ve
 
 - Update Go modules and GitHub Actions in [#840](https://github.com/cybozu-go/cke/pull/840)
 - Update Sabakan mock in [#840](https://github.com/cybozu-go/cke/pull/840)
+- Disable deprecation check for Endpoints temporarily in [#840](https://github.com/cybozu-go/cke/pull/840)
 
 ## Ancient changes