+++ CHANGELOG.md
@@ -14,6 +14,7 @@ This project employs a versioning scheme described in [RELEASE.md](RELEASE.md#ve
 - Disable deprecation check for Endpoints temporarily in [#840](https://github.com/cybozu-go/cke/pull/840)
 - Update Ubuntu of CKE container base to 24.04 in [#841](https://github.com/cybozu-go/cke/pull/841)
 - Update miscellaneous tools in [#841](https://github.com/cybozu-go/cke/pull/841)
+- Update containerd for mtest to 2.2.1 in [#841](https://github.com/cybozu-go/cke/pull/841)
 
 ## Ancient changes