+++ CHANGELOG.md
@@ -15,6 +15,7 @@ This project employs a versioning scheme described in [RELEASE.md](RELEASE.md#ve
 - Update Ubuntu of CKE container base to 24.04 in [#841](https://github.com/cybozu-go/cke/pull/841)
 - Update miscellaneous tools in [#841](https://github.com/cybozu-go/cke/pull/841)
 - Update containerd for mtest to 2.2.1 in [#841](https://github.com/cybozu-go/cke/pull/841)
+- Enable coordinated leader election by default in [#841](https://github.com/cybozu-go/cke/pull/841)
 
 ## Ancient changes