+++ CHANGELOG.md
@@ -16,6 +16,7 @@ This project employs a versioning scheme described in [RELEASE.md](RELEASE.md#ve
 - Update miscellaneous tools in [#841](https://github.com/cybozu-go/cke/pull/841)
 - Update containerd for mtest to 2.2.1 in [#841](https://github.com/cybozu-go/cke/pull/841)
 - Enable coordinated leader election by default in [#841](https://github.com/cybozu-go/cke/pull/841)
+- Update ClusterRole `system:kube-apiserver-to-kubelet` for kubelet fine-grained authorization in [#841](https://github.com/cybozu-go/cke/pull/841)
 
 ## Ancient changes