+++ CHANGELOG.md
@@ -6,6 +6,8 @@ This project employs a versioning scheme described in [RELEASE.md](RELEASE.md#ve
 
 ## [Unreleased]
 
+## [1.33.0-rc.1]
+
 ### Changed
 
 - Support Kubernetes 1.33 in [#841](https://github.com/cybozu-go/cke/pull/841)
@@ -42,4 +44,5 @@ This project employs a versioning scheme described in [RELEASE.md](RELEASE.md#ve
 - See [release-1.13/CHANGELOG.md](https://github.com/cybozu-go/cke/blob/release-1.13/CHANGELOG.md) for changes in CKE 1.13.
 - See [release-1.12/CHANGELOG.md](https://github.com/cybozu-go/cke/blob/release-1.12/CHANGELOG.md) for changes in CKE 1.12.
 
-[Unreleased]: https://github.com/cybozu-go/cke/compare/v1.32.6...HEAD
+[Unreleased]: https://github.com/cybozu-go/cke/compare/v1.33.0-rc.1...HEAD
+[1.33.0-rc.1]: https://github.com/cybozu-go/cke/compare/v1.32.6...v1.33.0-rc.1